+++ changelog.md
@@ -1,3 +1,11 @@
+# Version 6.0.1 - dd.07.2024
+
+## Bugfix
+
+- round light color of all rules in ``habapp_rules.actors.light_hcl`` to integer values to avoid strange formating in OpenHAB
+- added config parameter ``leaving_only_if_on`` to ``habapp_rules.actors.config.light.LightParameter`` to disable unexpected leaving light, if light was not on and leaving started
+- fixed bug in all shading rules of ``habapp_rules.actors.shading`` which did not switch to sleeping state if previous state was Auto_DoorOpen
+
 # Version 6.0.0 - 27.06.2024
 
 ## Breaking changes